Add feature to test Helm charts
Using chart testing (https://github.com/helm/chart-testing), add job to lint
Helm charts.

Chart testing is pretty extensive and we can easily add other tests using it.

IMHO this is the first step, next would be to remove the executor by adding `ct`
to architect docker image. After that look to extend architect command itself to
either call ct or use ct as a module (not sure if that is possible).