Options to disable and make helm lint less strict
There are cases where existing apps (not managed apps) will have a minimal Chart
file. In these cases, give the developers an option to make the check less
strict or completely disable.

Disabling might be useful where templates deliberately reference empty or non
existing values (in cases where values are taken from installations repo). It is
possible to create dummy values, but I think giving the option to disable
creates less of an impediment to development.

In both cases (lenient and disable), the default is false.