Use `charts` instead of `chart-dirs`
Using `charts` implies that version bump check is disabled, it is also better to
be precise as not all charts in a directory maybe templated correctly. For
example, `push-to-app-catalog` job only takes a single chart and will template
that.

The command can still support multiple charts if need be, but now we can
also specify a single chart.