jobs: add resource_class parameters to all jobs (#124)
* jobs: add resource_class parameters to all jobs

While switching to architect orb within Go modules work we discovered
that some codebases require more resources for particular jobs. The
problem we hit was aws-operator requiring more memory for building the
binary.

This change adds `resource_class` parameter to all (excluding legacy)
jobs. If not specified in CircleCI the default value is always "medium"
for both machine and docker executors. For some jobs the default is
changed to "small" to save some money.

- [x] src/jobs/changelog-lint.yaml - small by default.
- [x] src/jobs/gitleaks.yaml - small by default.
- [x] src/jobs/go-build.yaml - medium by default.
- [x] src/jobs/go-test.yaml - medium by default.
- [x] src/jobs/integration-test.yaml - medium by default.
- [x] src/jobs/push-to-app-catalog.yaml - small by default.
- [x] src/jobs/push-to-app-collection.yaml - small by default.
- [x] src/jobs/push-to-docker.yaml - small by default.
- [x] src/jobs/go-architect-legacy.yaml **SKIP** legacy job.
- [x] src/jobs/go-test-legacy.yaml **SKIP** legacy job.
- [x] src/jobs/push-to-docker-legacy.yaml **SKIP** legacy job.