Support Chart testing config file (#143)
* Support Chart testing config file

This is to support the use of additional Helm chart repositories.
Config file is needed for this option according to the docs
https://github.com/helm/chart-testing/tree/v2.4.1#using-private-chart-repositories

Additional repos are needed to satisfy Chart dependencies when testing. Maybe in
the future we can copy all dependencies to our own catalogs, then we won't need
to rely on external deps and we can just add our catalogs. But until that is
in place, we need a way to specify repos.

* Add missing config argument

* Update docs

Include usage of `ct_config`

* Update docs/run_kat_tests.md

Co-authored-by: Jose Armesto <github@armesto.net>

Co-authored-by: Jose Armesto <github@armesto.net>