An image visibility depends on the repo visibility
I've tried playing around with dynamic Circle-CI jobs and environment
variables, but nothing has really worked out. So I guess that the
easiest approach would be to check whether the repo is private or public
before pushing the image.

I've also added a parameter where all the registries that can be used
for private images should be listed. If a project is private, image
will be pushed only to registries listed under this parameter.

If one wants to bypass that check, there is an option to pass an
parameter `force-public`, that would just jump over the whole
visibility check. I haven't added `force-private`, because I'm not sure
about a use case. IF project is public, I don't think we need to force a
private image, because anyway anybody should be able to build it
themselves.